before merge patch of gemmini dialect

+++ midend/lib/Dialect/Gemmini/Transforms/LegalizeForLLVMExport.cpp
@@ -181,12 +181,12 @@ struct GemminiConfigNormOpLowering : public ConvertOpToLLVMPattern<ConfigNormOp>
     Location loc = configNormOp.getLoc();
     // ROCC_INSTRUCTION_RS1_RS2(XCUSTOM_ACC, 
     // (((uint64_t) ((uint32_t) q_const)) << 32) | ((q_const_type & 1) << 18) | ((set_stats_id_only & 1) << 17) | ((act_msb & 1) << 16) | ((uint64_t)stat_id << 8) | CONFIG_BERT, ((uint64_t)((uint32_t)(igelu_qc)) << 32) | ((uint64_t)((uint32_t)(igelu_qb))), k_CONFIG)
-    uint64_t rs1 = (((uint64_t) (configNormOp.getQConst())) << 32) |
+    uint64_t rs1 = (((uint64_t) ((uint32_t)configNormOp.getQConst())) << 32) |
                    (configNormOp.getQConstType() & 1) << 18 |
                    (configNormOp.getSetStatsIdOnly() & 1) << 17 |
                    (configNormOp.getActMsb() & 1) << 16 |
                    configNormOp.getStatsId() << 8 | CONFIG_BERT;
-    uint64_t rs2 = (((uint64_t) configNormOp.getIguluQc()) << 32) | ((uint64_t) ((uint32_t)configNormOp.getIguluQb()));
+    uint64_t rs2 = (((uint64_t) ((uint32_t)configNormOp.getIguluQc())) << 32) | ((uint64_t) ((uint32_t)configNormOp.getIguluQb()));
     Value rs1Value = rewriter.create<arith::ConstantOp>(
         loc, rewriter.getI64IntegerAttr(rs1));
     Value rs2Value = rewriter.create<arith::ConstantOp>(
